requirement.txt
google-genai
python-dotenv
markdown
